requirements.txt
Flask==2.3.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.2.1
WTForms==3.1.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
werkzeug==2.3.0
email-validator==2.1.0
Pillow==10.1.0  # For image handling
# Add these
openai==0.28.0
langchain==0.1.0
